+++ web/components/author-summary.tmpl.html
@@ -1,37 +1,37 @@
 <script defer>
   customElements.define(
-    "anthology-summary",
+    "author-summary",
     class extends HTMLElement {
       static get observedAttributes() {
-        return ["data-anthology-id"];
+        return ["data-author-id"];
       }
 
       constructor() {
         super();
 
-        const template = document.getElementById("anthology-template").content;
+        const template = document.getElementById("author-template").content;
         const shadowRoot = this.attachShadow({ mode: "open" }).appendChild(
           template.cloneNode(true)
         );
       }
 
-      get anthologyId() {
-        const idAttribute = this.attributes.getNamedItem("data-anthology-id");
+      get authorId() {
+        const idAttribute = this.attributes.getNamedItem("data-author-id");
         return idAttribute === null ? null : idAttribute.value;
       }
 
       get link() {
-        return this.shadowRoot.querySelector("anthology-link");
+        return this.shadowRoot.querySelector("author-link");
       }
 
       attributeChangedCallback() {
-        this.link.setAttribute("data-anthology-id", this.anthologyId);
+        this.link.setAttribute("data-author-id", this.authorId);
       }
     }
   );
 </script>
 
-<template id="anthology-template">
+<template id="author-template">
   <style>
     ::slotted(li) {
       list-style: none;
@@ -43,7 +43,7 @@
   </style>
 
   <section>
-    <anthology-link><slot name="title"></slot></anthology-link>
+    <author-link><slot name="title"></slot></author-link>
     <ul>
       <slot></slot>
     </ul>
